[bugfix] pad `digest` with 1s
A decoder must be able to determine if it has decoded all elements of GCS.
And `N` cannot be used for the purpose, since duplicates in `hash-values` are not encoded in GCS (see step 8-3).
Fortunately, each element encoded in GCS is guaranteed to have at least one zero-bit (see step 8-7).
Therefore, if the encoder uses 1s as a padding, the decoder can determine the end of the GCS (much like is done in HPACK).