Few minor tweaks
Just a couple copy edits, otherwise it looks good with a comment...

In the good case, where `trailer-update` is indicated in the header `Cache-Control` but no subsequent trailing `Cache-Control` is provided:

```http-message
HTTP/1.1 200 OK
Content-Type: text/html
Cache-Control: max-age=3600, trailer-update

[body]
```

It says that "Caches that do not implement this specification will cache it by the header policy; caches that do implement will see no updates in the trailer and do the same.", but that means that the `trailer-update` directive does get stored and forwarded on by the cache (since it is never removed). I'm just wanting to make sure there's no unintended down stream consequence from that. We may want to explicitly state that the presence of a `trailer-update`  directive does not *mandate* that a `Cache-Control`  trailer is present.

The only other thing that I can think of is ... when a trailing `Cache-Control` directive is received and it contains an updated timestamp (max-stale, expires, etc) is that calculated from is there any impact on the origin time from which that is calculated. Let's take an extreme example: I start streaming a resource back from the server with a max-age of 30 seconds. 20 seconds later the resource is done streaming and I update the `Cache-Control` with a max-age of 10 seconds.... Is the resource already stale (age calculated from the time the resource started to be received) or is the age calculated from the time the `Cache-Control` was updated and the resource was committed to the cache? Or does that not even matter here.